// src/api/enums/EscrowMessageType.ts
// Copyright (c) 2017-2018, The Particl Market developers
// Distributed under the GPL software license, see the accompanying
// file COPYING or https://github.com/particl/particl-market/blob/develop/LICENSE

/**
 * EscrowMessageType
 *
 */

export enum EscrowMessageType {

    MPA_LOCK = 'MPA_LOCK',
    MPA_REQUEST_REFUND = 'MPA_REQUEST_REFUND',
    MPA_REFUND = 'MPA_REFUND',
    MPA_RELEASE = 'MPA_RELEASE'

}